Allow to choose an examiner as ruler and finish signal
Instead of choosing an examiner at random to get the line addresses (aka
the ruler), let the user to choose one (the first one by default).

Use the ruler to know when to stop yielding lines: once the ruler gets
exhausted that's all, no matter if the rest of the examiners are
exhausted or not, the line generation ends.

This fixes some corner cases when the data was multiple of the line size
and the code raised StopIteration errors.

If one examiner gets exhausted and *after that* the ruler does not get
exhausted, it is considered an error. If we find a valid case of use for
that, we can relax the check with a parameter (like "keep" or something
like that).